changed Item.damage from double to int

+++ src/main/java/LairOfTheForgottenBeast/domain/prop/Item.java
@@ -7,7 +7,7 @@ public class Item implements Prop
    String shortDescription; // This is read as the default description
    String longDescription; // This is read someone "examines" a prop object
    String properties; // Stone, Wood, leather, metal, etc.
-   double damage = 1.0;
+   int damage = 1.0;
 
    public Item() { }
